requirements.txt
Pillow==9.5.0
pilmoji==2.0.2
py3-pinterest
selenium-wire